+++ entityProject.vcxproj
@@ -19,62 +19,59 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="src\CleanUpGroup.cpp" />
-    <ClCompile Include="src\EntityPool.cpp" />
-    <ClCompile Include="src\InitializationGroup.cpp" />
-    <ClCompile Include="src\JobManager.cpp" />
-    <ClCompile Include="src\JobWorker.cpp" />
-    <ClCompile Include="src\main.cpp" />
-    <ClCompile Include="src\PresentationGroup.cpp" />
-    <ClCompile Include="src\Schedule.cpp" />
-    <ClCompile Include="src\SimulationGroup.cpp" />
-    <ClCompile Include="src\SystemGroup.cpp" />
-    <ClCompile Include="src\TaskQueue.cpp" />
+    <ClCompile Include="examples\main.cpp" />
+    <ClCompile Include="src\ECS\EntityPool.cpp" />
+    <ClCompile Include="src\ECS\JobManager.cpp" />
+    <ClCompile Include="src\System\CleanUpGroup.cpp" />
+    <ClCompile Include="src\System\InitializationGroup.cpp" />
+    <ClCompile Include="src\System\PresentationGroup.cpp" />
+    <ClCompile Include="src\System\SystemScheduler.cpp" />
+    <ClCompile Include="src\System\SimulationGroup.cpp" />
+    <ClCompile Include="src\System\SystemGroup.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="include\CallbackListMT.hpp" />
-    <ClInclude Include="include\CallBackListST.hpp" />
-    <ClInclude Include="include\CleanUpGroup.h" />
-    <ClInclude Include="include\ComponentPoolManager.hpp" />
-    <ClInclude Include="include\ecsTest.h" />
-    <ClInclude Include="include\Entity.h" />
-    <ClInclude Include="include\EntityEventMixin.hpp" />
-    <ClInclude Include="include\EntityPool.h" />
-    <ClInclude Include="include\Event.hpp" />
-    <ClInclude Include="include\EventDispatcherMT.hpp" />
-    <ClInclude Include="include\EventDispatcherST.hpp" />
-    <ClInclude Include="include\EventQueue.hpp" />
-    <ClInclude Include="include\IEvent.hpp" />
-    <ClInclude Include="include\group.hpp" />
-    <ClInclude Include="include\HashFunctions.hpp" />
-    <ClInclude Include="include\HopscotchHashMap.h" />
-    <ClInclude Include="include\InitializationGroup.h" />
-    <ClInclude Include="include\intrusive_ptr.h" />
-    <ClInclude Include="include\JobBarrier.h" />
-    <ClInclude Include="include\JobDebugger.h" />
-    <ClInclude Include="include\JobDeque.hpp" />
-    <ClInclude Include="include\JobHandle.hpp" />
-    <ClInclude Include="include\JobManager.h" />
-    <ClInclude Include="include\JobRecorder.h" />
-    <ClInclude Include="include\JobWorker.h" />
-    <ClInclude Include="include\OrderedQueueList.hpp" />
-    <ClInclude Include="include\PreInitializationGroup.h" />
-    <ClInclude Include="include\PresentationGroup.h" />
-    <ClInclude Include="include\Schedule.h" />
-    <ClInclude Include="include\Signal.hpp" />
-    <ClInclude Include="include\SimulationGroup.h" />
-    <ClInclude Include="include\SparseHopscotch.h" />
-    <ClInclude Include="include\SparseSet.h" />
-    <ClInclude Include="include\Storage.hpp" />
-    <ClInclude Include="include\StorageMixin.hpp" />
-    <ClInclude Include="include\SystemBase.hpp" />
-    <ClInclude Include="include\SystemGroup.h" />
-    <ClInclude Include="include\taskPtr.hpp" />
-    <ClInclude Include="include\TaskQueue.h" />
+    <ClInclude Include="examples\ecsTest.h" />
+    <ClInclude Include="include\Engine\Core\Containers\HashFunctions.hpp" />
+    <ClInclude Include="include\Engine\Core\Containers\HopscotchHashMap.h" />
+    <ClInclude Include="include\Engine\Core\Containers\SparseHopscotch.h" />
+    <ClInclude Include="include\Engine\Core\Experimental\intrusive_ptr.h" />
+    <ClInclude Include="include\Engine\Core\Experimental\JobDeque.hpp" />
+    <ClInclude Include="include\Engine\Core\Experimental\ThreadSafeQueue.h" />
+    <ClInclude Include="include\Engine\Core\JobBarrier.h" />
+    <ClInclude Include="include\Engine\Core\JobDebugger.h" />
+    <ClInclude Include="include\Engine\Core\JobHandle.hpp" />
+    <ClInclude Include="include\Engine\Core\JobManager.h" />
+    <ClInclude Include="include\Engine\Core\JobRecorder.h" />
+    <ClInclude Include="include\Engine\Core\JobWorker.h" />
+    <ClInclude Include="include\Engine\Core\taskPtr.hpp" />
+    <ClInclude Include="include\Engine\ECS\ComponentPoolManager.hpp" />
+    <ClInclude Include="include\Engine\ECS\Entity.h" />
+    <ClInclude Include="include\Engine\ECS\EntityEventMixin.hpp" />
+    <ClInclude Include="include\Engine\ECS\EntityPool.h" />
+    <ClInclude Include="include\Engine\ECS\Events\CallbackListMT.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\CallBackListST.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\Event.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\EventDispatcherMT.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\EventDispatcherST.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\EventQueue.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\IEvent.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\OrderedQueueList.hpp" />
+    <ClInclude Include="include\Engine\ECS\Events\Signal.hpp" />
+    <ClInclude Include="include\Engine\ECS\group.hpp" />
+    <ClInclude Include="include\Engine\ECS\SparseSet.h" />
+    <ClInclude Include="include\Engine\ECS\Storage.hpp" />
+    <ClInclude Include="include\Engine\ECS\StorageMixin.hpp" />
+    <ClInclude Include="include\Engine\ECS\System\CleanUpGroup.h" />
+    <ClInclude Include="include\Engine\ECS\System\InitializationGroup.h" />
+    <ClInclude Include="include\Engine\ECS\System\PreInitializationGroup.h" />
+    <ClInclude Include="include\Engine\ECS\System\PresentationGroup.h" />
+    <ClInclude Include="include\Engine\ECS\System\SystemScheduler.h" />
+    <ClInclude Include="include\Engine\ECS\System\SimulationGroup.h" />
+    <ClInclude Include="include\Engine\ECS\System\SystemBase.hpp" />
+    <ClInclude Include="include\Engine\ECS\System\SystemGroup.h" />
+    <ClInclude Include="include\Engine\ECS\typeList.hpp" />
+    <ClInclude Include="include\Engine\ECS\World.h" />
     <ClInclude Include="include\TestFramework.hpp" />
-    <ClInclude Include="include\ThreadSafeQueue.h" />
-    <ClInclude Include="include\typeList.hpp" />
-    <ClInclude Include="include\World.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>16.0</VCProjectVersion>
