+++ include/Engine/Core/JobDebugger.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "TestFramework.hpp"
+#include <optional>
 
 namespace ECS::JobSystem::Debug {
