+++ include/Engine/ECS/ComponentPoolManager.hpp
@@ -5,7 +5,7 @@
 #include <vector>
 #include <memory>
 #include <bitset>
-#include "HopscotchHashMap.h"
+#include "Engine/Core/Containers/HopscotchHashMap.h"
 
 namespace ECS{
