+++ include/Engine/ECS/EntityEventMixin.hpp
@@ -1,23 +1,18 @@
-#ifndef ECS_STORAGEMIXIN_HPP
-#define ECS_STORAGEMIXIN_HPP
-#include "CallbackListST.hpp"
-#include "Entity.h"
+#ifndef ECS_ENTITYEVENTMIXIN_HPP
+#define ECS_ENTITYEVENTMIXIN_HPP
+
+#include "Engine/ECS/Events/CallBackListST.hpp"
 
 namespace ECS{
 
-namespace COMPONENT{
-
-    // Event 向けミックスイン
-    struct EventsMixin {
-
+    template<typename Derived>
+    struct EntityEventMixin {
         // コールバックの登録
         auto& on_construct() { return construct; }
-        auto& on_update() { return update; }
         auto& on_destroy() { return destroy; }
 
-        // 利用側（派生クラス）から呼び出す通知用メソッド
+        // 利用側から呼び出す通知用メソッド
         void notify_construct(EntityID entity) { construct(entity); }
-        void notify_update(EntityID entity) { update(entity); }
         void notify_destroy(EntityID entity) { destroy(entity); }
 
     private:
@@ -25,13 +20,10 @@ namespace COMPONENT{
 
     private:
         EVENT::CallbackList_Single<void(const EntityID)> construct;
-        EVENT::CallbackList_Single<void(const EntityID)> update;
         EVENT::CallbackList_Single<void(const EntityID)> destroy;
     };
 
-}//namespace COMPONENT_STORAGE
+}// namespace ECS
 
-}//namespace ECS
-
-#endif // !ECS_STORAGEMIXIN_HPP
+#endif // !ECS_ENTITYEVENTMIXIN_HPP
