+++ include/Engine/ECS/Events/Event.hpp
@@ -8,7 +8,7 @@
 #include <utility>
 #include <stdexcept>
 #include <cassert>
-#include "HashFunctions.hpp"
+#include "Engine/Core/Containers/HashFunctions.hpp"
 
 namespace ECS{
