+++ include/Engine/ECS/SparseSet.h
@@ -6,7 +6,7 @@
 #include <vector>
 #include <array>
 #include <limits>
-#include "HashFunctions.hpp"
+#include "Engine/Core/Containers/HashFunctions.hpp"
 #include "TestFramework.hpp"
 
 namespace ECS{
