+++ include/Engine/ECS/StorageMixin.hpp
@@ -1,18 +1,23 @@
-#ifndef ECS_ENTITYEVENTMIXIN_HPP
-#define ECS_ENTITYEVENTMIXIN_HPP
-
-#include "CallBackListST.hpp"
+#ifndef ECS_STORAGEMIXIN_HPP
+#define ECS_STORAGEMIXIN_HPP
+#include "Engine\ECS\Events\CallBackListST.hpp"
+#include "Entity.h"
 
 namespace ECS{
 
-    template<typename Derived>
-    struct EntityEventMixin {
+namespace COMPONENT{
+
+    // Event 向けミックスイン
+    struct EventsMixin {
+
         // コールバックの登録
         auto& on_construct() { return construct; }
+        auto& on_update() { return update; }
         auto& on_destroy() { return destroy; }
 
-        // 利用側から呼び出す通知用メソッド
+        // 利用側（派生クラス）から呼び出す通知用メソッド
         void notify_construct(EntityID entity) { construct(entity); }
+        void notify_update(EntityID entity) { update(entity); }
         void notify_destroy(EntityID entity) { destroy(entity); }
 
     private:
@@ -20,10 +25,13 @@ namespace ECS{
 
     private:
         EVENT::CallbackList_Single<void(const EntityID)> construct;
+        EVENT::CallbackList_Single<void(const EntityID)> update;
         EVENT::CallbackList_Single<void(const EntityID)> destroy;
     };
 
-}// namespace ECS
+}//namespace COMPONENT_STORAGE
 
-#endif // !ECS_ENTITYEVENTMIXIN_HPP
+}//namespace ECS
+
+#endif // !ECS_STORAGEMIXIN_HPP
