--- src/JobWorker.cpp
@@ -1,2 +1,0 @@
-#include "JobWorker.h"
-
