--- src/TaskQueue.cpp
@@ -1,5 +1,0 @@
-#include "TaskQueue.h"
-#include "JobManager.h"
-
-namespace ECS::JobSystem{
-}
